s21_Python/Python_Bootcamp.Team1/src/requirements.txt
python-dotenv==1.0.1
SQLAlchemy==2.0.32
aiogram
sphinx
psycopg2-binary